requirements.txt
apache-airflow==1.9.0
gym==0.15.4
stable_baselines==2.9.0
PyYAML==5.3
Flask==0.11.1
flask_restplus==0.12.1
tensorflow==1.13.1